Create Bioinformatics
This code defines a generic Node class, which has a Data property and a Next property. The Data property stores the data of the Node, and the Next property points to the next Node in the LinkedList.
The LinkedList class has a head property, which points to the first Node in the LinkedList. It also has two methods:
1 Add: This method creates a new Node with the provided data and adds it to the end of the LinkedList.
2 DisplayNodes: This method iterates over each Node in the LinkedList and prints out its data.